update tasks & users

+++ app/Http/Controllers/TasksController.php
@@ -18,11 +18,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        };
+
+        return view('tasks.index', $data);
     }
 
     /**
@@ -33,10 +41,12 @@ class TasksController extends Controller
     public function create()
     {
         $task = new Task;
-
-        return view('tasks.create', [
-            'task' => $task,
-        ]);
+        if (\Auth::check()) {
+            return view('tasks.create', [
+                'task' => $task,
+            ]);
+        };
+        return redirect('/');
     }
 
     /**
@@ -50,10 +60,11 @@ class TasksController extends Controller
         $this->validate($request, [
             'status' => 'required',
         ]);
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
         return redirect('/');
     }
@@ -67,9 +78,12 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::user()->id === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        };
+        return redirect('/');
     }
 
     /**
@@ -81,10 +95,12 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
-
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::user()->id === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        };
+        return redirect('/');
     }
 
     /**
@@ -99,7 +115,7 @@ class TasksController extends Controller
         $this->validate($request, [
             'status' => 'required',
         ]);
-        $task = Task::find($id);
+        $task = $request->user()->tasks()->find($id);
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
@@ -116,7 +132,9 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        };
 
         return redirect('/');
     }
